[INLONG-7614][Sort] Fix sort pulsar connector loss data when using admin url
